build: Создать маршрутизацию
