+++ src/blocks/page/_loading/page_loading.css
@@ -1,0 +1,6 @@
+.page_loading {
+    min-height: 100vh;
+    display: flex;
+    align-items: center;
+    justify-content: center;
+}
