+++ src/components/Login.js
@@ -1,0 +1,7 @@
+import React from 'react';
+
+function Login() {
+  return <main className="content"></main>;
+}
+
+export default Login;
